new README.md proposal
Clearly direct readers to the main IO page where they can easily find
more information on all aspects of Quarks.  Right now its more work than
it should be for someone to be able to locate the releases/downloads,
faq, ....  The link to IO at the top of the page easily goes unnoticed -
I know! :-)

Remove all of the overview info as it's already covered in the main IO
page and its links.  Avoid maintaining multiple versions of it.